Update README.md
Game Storyline 
In Mushroom Kingdom Rivals, a racer wants to win the big Mushroom Kingdom race. But bad guys led by Bowser try to mess it up. The racer teams up with friends, stops Bowser, saves Princess Peach, and wins the race to bring back peace.

Game Story
The story of the game Kart Racer is that first of all there are different tracks available and user has to pick one of the tracks in which the user does racing with their opponents. Lots of really good racers from everywhere are joining in means there are total of 8 racers who are participating in the race. The tracks are so good, tracks are having some hurdles so it’s a great competition between the users and users earns points by collecting coins and collect different things like banana peel and mushrooms. There are 2 to 3 laps for each track. This is kart race and those who complete the laps early will win the game.